try to install homebrew if it's not there

+++ evremixes-downloader.sh
@@ -9,10 +9,10 @@ NC='\033[0m' # No Color
 temp_dir=$(mktemp -d -t evremixes-XXXXXX)
 
 # Check if ffmpeg is installed
-if ! command -v ffmpeg &> /dev/null; then
+if ! command -v ffmpeg &>/dev/null; then
     echo -e "${RED}Warning:${NC} ffmpeg is not installed."
     # Check if Homebrew is installed
-    if command -v brew &> /dev/null; then
+    if command -v brew &>/dev/null; then
         echo -e "${GREEN}Homebrew is installed, attempting to install ffmpeg...${NC}"
         brew install ffmpeg
         if [ $? -ne 0 ]; then
@@ -21,9 +21,42 @@ if ! command -v ffmpeg &> /dev/null; then
             exit 1
         fi
     else
-        echo -e "${RED}Error:${NC} Homebrew is not installed. Cannot install ffmpeg."
-        rm -r "${temp_dir}"
-        exit 1
+        echo -e "${RED}Warning:${NC} Homebrew is not installed. Attempting to install it now..."
+        # Install Homebrew
+        /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+        if [ $? -ne 0 ]; then
+            echo -e "${RED}Failed to install Homebrew.${NC}"
+            rm -r "${temp_dir}"
+            exit 1
+        fi
+
+        # Add brew to PATH for this session based on architecture and OS
+        case "$(uname -m)" in
+        'arm64')
+            # For ARM architecture
+            export PATH="/opt/homebrew/bin:$PATH"
+            ;;
+        'x86_64')
+            # For x86 architecture
+            case "$(uname -s)" in
+            'Linux')
+                export PATH="/home/linuxbrew/.linuxbrew/bin:$PATH"
+                ;;
+            'Darwin')
+                export PATH="/usr/local/bin:$PATH"
+                ;;
+            esac
+            ;;
+        esac
+
+        # Now install ffmpeg
+        echo -e "${GREEN}Homebrew installed successfully, attempting to install ffmpeg...${NC}"
+        brew install ffmpeg
+        if [ $? -ne 0 ]; then
+            echo -e "${RED}Failed to install ffmpeg.${NC}"
+            rm -r "${temp_dir}"
+            exit 1
+        fi
     fi
 fi
 
@@ -42,7 +75,7 @@ else
 fi
 
 # Check if curl is installed
-if ! command -v curl &> /dev/null; then
+if ! command -v curl &>/dev/null; then
     echo -e "${RED}Error:${NC} curl is not installed. Please install it and try again."
     exit 1
 fi
